# specs/openapi.json
{"openapi":"3.0.2","info":{"title":"Custom title","description":"This is a very custom OpenAPI schema","version":"2.5.0"},"paths":{"/healthz":{"get":{"summary":"Get Health","operationId":"get_health_healthz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/accounts/{account_id}":{"get":{"summary":"Read Account","operationId":"read_account_accounts__account_id__get","parameters":[{"required":true,"schema":{"title":"Account Id","type":"integer"},"name":"account_id","in":"path"},{"required":false,"schema":{"title":"Status Code","default":200},"name":"status_code","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Update Account","operationId":"update_account_accounts__account_id__put","parameters":[{"required":true,"schema":{"title":"Account Id","type":"integer"},"name":"account_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Remove Account","operationId":"remove_account_accounts__account_id__delete","parameters":[{"required":true,"schema":{"title":"Deleted","type":"boolean"},"name":"deleted","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/accounts":{"post":{"summary":"Create Account","operationId":"create_account_accounts_post","parameters":[{"required":true,"schema":{"title":"Account Id","type":"integer"},"name":"account_id","in":"query"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Account":{"title":"Account","required":["name","balance"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"description":{"title":"Description","type":"string"},"balance":{"title":"Balance","type":"number"},"active":{"title":"Active","type":"boolean","default":true}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}